React 版本默认实现 alert、confirm、jumpTo、updateLocation、isCurrentUrl、notify，降低上手成本
