+++ examples/components/Play.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import {toast} from '../../src/components/Toast';
 import {render} from '../../src/index';
+import {normalizeLink} from '../../src/utils/normalizeLink';
 import {alert, confirm} from '../../src/components/Alert';
 import axios from 'axios';
 import Frame from 'react-frame-component';
@@ -87,44 +88,6 @@ export default class PlayGround extends React.Component {
       style: {
         height: '100%'
       }
-    };
-    const normalizeLink = to => {
-      to = to || '';
-      const location = router.getCurrentLocation();
-
-      if (to && to[0] === '#') {
-        to = location.pathname + location.search + to;
-      } else if (to && to[0] === '?') {
-        to = location.pathname + to;
-      }
-
-      const idx = to.indexOf('?');
-      const idx2 = to.indexOf('#');
-      let pathname = ~idx
-        ? to.substring(0, idx)
-        : ~idx2
-        ? to.substring(0, idx2)
-        : to;
-      let search = ~idx ? to.substring(idx, ~idx2 ? idx2 : undefined) : '';
-      let hash = ~idx2 ? to.substring(idx2) : location.hash;
-
-      if (!pathname) {
-        pathname = location.pathname;
-      } else if (pathname[0] != '/' && !/^https?:\/\//.test(pathname)) {
-        let relativeBase = location.pathname;
-        const paths = relativeBase.split('/');
-        paths.pop();
-        let m;
-        while ((m = /^\.\.?\//.exec(pathname))) {
-          if (m[0] === '../') {
-            paths.pop();
-          }
-          pathname = pathname.substring(m[0].length);
-        }
-        pathname = paths.concat(pathname).join('/');
-      }
-
-      return pathname + search + hash;
     };
     this.env = {
       session: 'doc',
